moved ProfileContent to its own component
